2009-12-16 09:59 UTC+0100 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * harbour/src/common/hbdate.c     * casting for some more pedantic C/C++ compilers
  * harbour/include/hbinit.h
    * allow to force static C++ initialization as startup code in
      GCC and SunPRO builds by HB_STATIC_STARTUP macro

+++ harbour/src/common/hbdate.c
@@ -925,7 +925,7 @@ long hb_timeUTCOffset( void ) /* in seconds */
       timeinfo = *localtime( &current );
       local = mktime( &timeinfo );
 #endif
-      return difftime( local, utc ) + ( timeinfo.tm_isdst > 0 ? 3600 : 0 );
+      return ( long ) difftime( local, utc ) + ( timeinfo.tm_isdst > 0 ? 3600 : 0 );
    }
 #endif
 }
